export theme to npm package
