+++ scripts/postbuild.js
@@ -62,6 +62,8 @@ const main = async () => {
   copy('dist/prime.css', `dist/prime@${version}.css`);
   copy('dist/icons.woff2', `dist/icons@${version}.woff2`);
 
+  copy('theme.js', 'dist/theme.js')
+
   // Update paths for versioned resources
   {
     const file = read(`dist/prime@${version}.css`).replaceAll(
